Move library to root of repository
A popular library installation technique is to download the library via GitHub's Clone or download > Download ZIP and then use the Arduino IDE's Sketch > Include Library > Add .ZIP Library on the downloaded file. This requires the library to be in the root of the repository, not in a subfolder. If the library is not in the root of the repository this installation technique fails:

Specified folder/zip file does not contain a valid library

This is the standard repository structure used in all official Arduino libraries:
https://github.com/arduino-libraries

This move is also required if you wanted to add your library to the Arduino Library Manager index, which provides an even easier installation option.

It appears there was previously some intention to do this since the library.properties and keywords.txt files are already in the root but those files will only be recognized by the Arduino IDE if they are in the library's root folder, not only the repository's root.